implement the concern for current_user action
