+++ db/migrate/20240514065215_remove_quantity_from_cart_items.rb
@@ -1,0 +1,6 @@
+class RemoveQuantityFromCartItems < ActiveRecord::Migration[7.1]
+  def change
+    remove_column :cart_items, :quantity, :integer
+    remove_column :carts, :date_added, :datetime
+  end
+end
